requirements.txt
Django==1.7.6
Markdown==2.6.1
django-bootstrap3==5.1.1
django-datetime-widget==0.9.3
django-debug-toolbar==1.3.0
django-suit==0.2.12
djangorestframework==3.1.0
psycopg2==2.6
pytz==2014.10
sqlparse==0.1.14
wsgiref==0.1.2
